Added load balancing to the parallel max flow.
The parallel max flow algorithm now sends an active node from one thread
with multiple active nodes to a thread with no active nodes. This allows
for more threads than the degree of the stating vertex. The code does
sometimes produce the wrong result, and there is a rare case of double
free somewhere.