[MF] Fixed race conditions that lead to wrong results
Threads now request nodes from the first available queue, and it cleans
up properly if it can't set the appropriate flags. A part of the proper
cleanup is to set the receiver thread to itself such that nothing
happens.

Added a counter for the amount of available nodes. This fixes the
problem where excess[s] would be loaded, excess[t] changed by a
different thread, and finally excess[t] loaded in the test of the
outermost while loop in execute. This sequence could lead to wrong
results.

Added an extra check against invalid nodes, and use
compare_exchange_strong to set the isActive flag for a node.

All queue.push call has been changed to queue.emplace.